storydredge/requirements.txt
requests==2.31.0
bs4==0.0.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
tqdm==4.66.1 